backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
rembg
requests
google-generativeai
python-dotenv
onnxruntime